# 20.api/10.swagger/cocktail/openapi.yaml
openapi: 3.0.1
info:
  title: Cocktails API
  description: API to manage cocktail info.
  contact:
    email: okarians.302.dev@gmail.com
  version: 0.1.4
servers:
  - url: 'https://release.com/api/v1/'
    description: Release server
  - url: 'http://localhost/api/v1/'
    description: Local server
tags:
  - name: CreateCocktails
    description: Create API for cocktail info.
  - name: GetCocktails
    description: Get API for cocktail info.
  - name: UpdateCocktails
    description: Update API for cocktail info.
  - name: DeleteCocktails
    description: Delete API for cocktail info.
  - name: SearchCocktails
    description: Search API for cocktail info.

paths:
  /cocktails/{id}:
    get:
      tags:
        - GetCocktails
      summary: Get cocktail info by id.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Cocktail id.
          required: true
          schema:
            $ref: '#/components/schemas/CocktailId'
      responses:
        '200':
          $ref: '#/components/responses/CocktailGetResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
    put:
      tags:
        - UpdateCocktails
      summary: Update cocktail info by id.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Cocktail id.
          required: true
          schema:
            $ref: '#/components/schemas/CocktailId'
      requestBody:
        description: Update cocktail info request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailDetailModel'
      responses:
        '201':
          $ref: '#/components/responses/CocktailPutResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
    delete:
      tags:
        - DeleteCocktails
      summary: Delete cocktail info by id.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          description: Cocktail id.
          required: true
          schema:
            $ref: '#/components/schemas/CocktailId'
      responses:
        '204':
          $ref: '#/components/responses/CocktailDeleteResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
  /cocktails:
    get:
      tags:
        - GetCocktails
      summary: Get cocktail info list.
      security:
        - bearerAuth: []
      responses:
        '200':
          $ref: '#/components/responses/CocktailListGetResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
    post:
      tags:
        - CreateCocktails
      summary: Create cocktail info.
      security:
        - bearerAuth: []
      requestBody:
        description: Create cocktail info request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailDetailModel'
      responses:
        '201':
          $ref: '#/components/responses/CocktailPostResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
  /cocktails/bulk_create:
    post:
      tags:
        - CreateCocktails
      summary: Create maltiple cocktail info.
      security:
        - bearerAuth: []
      requestBody:
        description: Delete cocktail info request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailList'
      responses:
        '201':
          $ref: '#/components/responses/CocktailPostResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
  /cocktails/bulk_update:
    post:
      tags:
        - UpdateCocktails
      summary: Update maltiple cocktail info.
      security:
        - bearerAuth: []
      requestBody:
        description: Delete cocktail info request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailList'
      responses:
        '201':
          $ref: '#/components/responses/CocktailPostResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
  /cocktails/bulk_delete:
    post:
      tags:
        - DeleteCocktails
      summary: Delete maltiple cocktail info.
      security:
        - bearerAuth: []
      requestBody:
        description: Delete cocktail info request body.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CocktailIdList'
      responses:
        '204':
          $ref: '#/components/responses/CocktailDeleteResponce'
        '400':
          $ref: '#/components/responses/400Model'
        '401':
          $ref: '#/components/responses/401Model'
        '409':
          $ref: '#/components/responses/409Model'
        '500':
          $ref: '#/components/responses/500Model'
components:
  responses:
    CocktailGetResponce:
      description: Get cocktail info response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CocktailModel'
    CocktailPostResponce:
      description: Create cocktail info response.
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
    CocktailPutResponce:
      description: Update cocktail info response.
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
    CocktailDeleteResponce:
      description: Delete cocktail info response.
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
    CocktailListGetResponce:
      description: Get cocktail info list response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CocktailList'
    400Model:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
          example:
            reason: Validation_Error
            msg: リクエストの形式が不正です
    401Model:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
          example:
            reason: Authentication_ERROR
            msg: 認証に失敗しました 
    409Model:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
          example:
            reason: DATA_INCONSITENCY
            msg: データ不整合のため、突合情報検索に失敗しました
    500Model:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '../components/CommonMessageModel.yaml'
          example:
            reason: SERVER_ERROR
            msg: サーバーエラー
  schemas:
    CocktailModel:
      description: Cocktail info model.
      type: object
      properties:
        cocktailId:
          $ref: '#/components/schemas/CocktailId'
        cocktailName:
          $ref: '#/components/schemas/CocktailName'
    CocktailList:
      description: Cocktail info list.
      type: array
      items:
        $ref: '#/components/schemas/CocktailModel'
    CocktailDetailModel:
      properties:
        cocktailId:
          $ref: '#/components/schemas/CocktailId'
        cocktailName:
          $ref: '#/components/schemas/CocktailName'
        materialIdList:
          description: Material id list.
          type: array
          items:
            type: integer
            description: Material id.
            example: 1
    CocktailIdList:
      description: Cocktail id list.
      type: array
      items:
        $ref: '#/components/schemas/CocktailId'

    # 以下、複数モデルで使用する共通部品
    CocktailId:
      description: Cocktail id
      type: integer
      format: integer
      example: 1
    CocktailName:
      description: cocktail name.
      type: string
      format: string
      example: ジントニック
  
security:
  - MapManagerAuth: [] 