# 20.api/20.source/20.cocktail/mycocktails.api.cocktailApi/Controllers/GetCocktailsApi.cs
/*
 * Cocktails API
 *
 * API to manage cocktail info.
 *
 * The version of the OpenAPI document: 0.1.4
 * Contact: okarians.302.dev@gmail.com
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Mvc;
using Newtonsoft.Json;
using mycocktails.api.cocktailApi.Attributes;
using mycocktails.api.cocktailApi.Models;
using mycocktails.library.cocktailApi.Controllers;
using mycocktails.library.cocktailApi.Models;
using mycocktails.library.common.Models;
using CommonMessageModel = mycocktails.library.common.Models.CommonMessageModel;
using System.Net;
using mycocktails.api.cocktailApi.Logics.intarfaces;
using Microsoft.Extensions.Logging;

namespace mycocktails.api.cocktailApi.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class GetCocktailController : GetCocktailsApiController
    {
        private readonly IGetCocktailLogic logic;
        private readonly ILogger<GetCocktailController> logger;

        /// <summary>
        /// Constructor.
        /// </summary>
        /// <param name="logic">logic</param>
        /// <param name="logger">logger</param>
        public GetCocktailController(
            IGetCocktailLogic logic,
            ILogger<GetCocktailController> logger)
        {
            this.logic = logic;
            this.logger = logger;
        }

        /// <summary>
        /// Get cocktail info by id.
        /// </summary>
        /// <param name="id">Cocktail id.</param>
        /// <response code="200">Get cocktail info response.</response>
        /// <response code="400">Bad Request</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="409">Conflict</response>
        /// <response code="500">Internal Server Error</response>
        [HttpGet]
        [Route("/api/v1/cocktails/{id}")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(CocktailModel))]
        [ProducesResponseType(statusCode: 400, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 401, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 409, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 500, type: typeof(CommonMessageModel))]
        public override IActionResult CocktailsIdGet([FromRoute][Required] int id)
        {
            ApiResponse result;

            try
            {
                result = logic.GetCocktail(id);
            }

            //
            catch (Exception ex)
            {
                CommonMessageModel error = new CommonMessageModel
                {
                    Msg = "An unknown error has occurred."
                };
                this.logger.LogError($"{error.Msg}");
                result = new ErrorResponse<CommonMessageModel>(HttpStatusCode.InternalServerError, error, ex.InnerException?.Message ?? ex.Message);
            }

            this.logger.LogInformation($"StatusCode: {result.StatusCode}");
            if (result.Success)
            {
                ApiResponse<CocktailModel> successResponse = (ApiResponse<CocktailModel>)result;
                return StatusCode((int)successResponse.StatusCode, successResponse.ResponseModel);
            }
            else
            {
                ApiResponse<CommonMessageModel> failureResponse = (ApiResponse<CommonMessageModel>)result;
                return StatusCode((int)failureResponse.StatusCode, failureResponse.ResponseModel);
            }
        }

        /// <summary>
        /// Get cocktail info list.
        /// </summary>
        /// <response code="200">Get cocktail info list response.</response>
        /// <response code="400">Bad Request</response>
        /// <response code="401">Unauthorized</response>
        /// <response code="409">Conflict</response>
        /// <response code="500">Internal Server Error</response>
        [HttpGet]
        [Route("/api/v1/cocktails")]
        [ValidateModelState]
        [ProducesResponseType(statusCode: 200, type: typeof(List<CocktailModel>))]
        [ProducesResponseType(statusCode: 400, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 401, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 409, type: typeof(CommonMessageModel))]
        [ProducesResponseType(statusCode: 500, type: typeof(CommonMessageModel))]
        public override IActionResult CocktailsGet()
        {
            ApiResponse result;

            try
            {
                result = logic.GetCocktailList();
            }

            //
            catch (Exception ex)
            {
                CommonMessageModel error = new CommonMessageModel
                {
                    Msg = "An unknown error has occurred."
                };
                this.logger.LogError($"{error.Msg}");
                result = new ErrorResponse<CommonMessageModel>(HttpStatusCode.InternalServerError, error, ex.InnerException?.Message ?? ex.Message);
            }

            this.logger.LogInformation($"StatusCode: {result.StatusCode}");
            if (result.Success)
            {
                ApiResponse<List<CocktailModel>> successResponse = (ApiResponse<List<CocktailModel>>)result;
                return StatusCode((int)successResponse.StatusCode, successResponse.ResponseModel);
            }
            else
            {
                ApiResponse<CommonMessageModel> failureResponse = (ApiResponse<CommonMessageModel>)result;
                return StatusCode((int)failureResponse.StatusCode, failureResponse.ResponseModel);
            }
        }
    }
}